Move common functionality to a form component.
